Pack progress results in C struct / ArrayBuffer to avoid json

+++ native/VoyageCrewRanker.h
@@ -25,7 +25,7 @@ struct RankedCrew {
 struct VoyageEstimate {
 	size_t primarySkill;
 	size_t secondarySkill;
-	double estimate;
+	float estimate;
 	std::array<Crew, SLOT_COUNT> crew;
 };
